LayoutTests/imported/w3c/web-platform-tests/credential-management/federatedcredential-framed-get.sub.https-expected.txt

CONSOLE MESSAGE: Error: assert_implements: `FederatedCredential` is supported. undefined

Harness Error (FAIL), message = Error: assert_implements: `FederatedCredential` is supported. undefined
